SQL/SQL/requirements.txt.txt
pandas
streamlit
flask
gevent
llama-cpp-agent
llama-cpp-python